incorporate Definition::is_shareable in Subgraphs
